requirements.txt
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
instagrapi==2.0.0
Pillow==9.5.0
pycodestyle==2.11.0
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
